move configuration to own file
